Split pages into per-model files
