+++ itests/pages/exceptions.py
@@ -1,0 +1,2 @@
+class NoSuchElementException(Exception):
+    pass
